+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next version
+
+* [yao #17](sdss/yao#17): Deal with case when the returned buffer does not match the expected size.
+
+
 ## 0.14.3 - January 12, 2024
 
 ### 🔧 Fixed